refactor(brett): extract sessions.ts (codes, admin token, idle/grace)
